updated code and added test files
