+++ src/components/EventMultipleComponent/index.ts
@@ -1,0 +1,1 @@
+export { KdEventMultipleComponent } from './eventMultipleComponent';
