+++ src/components/eventSingleComponent/eventSingleComponent.scss
@@ -1,1 +1,97 @@
-@use '~@kyndryl-design-system/shidoka-foundation/scss/utility/layout.scss';
+/**
+ * Copyright Kyndryl, Inc. 2022
+ */
+
+ @use 'sass:map';
+
+ @use '~@kyndryl-design-system/shidoka-foundation/scss/variables/breakpoints.scss';
+ @use '~@kyndryl-design-system/shidoka-foundation/scss/utility/grid.scss';
+ @use '~@kyndryl-design-system/shidoka-foundation/scss/utility/typography.scss';
+ @use '~@kyndryl-design-system/shidoka-foundation/scss/utility/layout.scss';
+ 
+ .kd-event-single-component {
+   background: var(--kd-color-background-ui-default);
+   color: var(--kd-color-text-primary);
+ 
+   h3 {
+     margin: 0 0 24px 0;
+   }
+ 
+   .kd-grid-padding {
+     padding: 0 1rem;
+   }
+ 
+   .full-bleed-grid {
+     display: grid;
+     grid-template-columns: 1fr 1fr;
+   }
+ 
+   .container-one {
+     grid-row-start: 1;
+     grid-column: 1;
+   }
+ 
+   .container-two {
+     grid-row-start: 1;
+     grid-column: 2;
+   }
+ 
+   //@include breakpoint(lg, xlg, max) {
+   @media (min-width: breakpoints.$bp-lg) {
+     .full-bleed-internal-container {
+       grid-column-start: 1;
+       grid-row-start: 1;
+       grid-column: 1 / span 2;
+     }
+   }
+ 
+   @media (min-width: breakpoints.$bp-lg) {
+     .container-two .internal-text-slot,
+     .container-two .internal-media-slot {
+       grid-column-start: 7;
+     }
+   }
+ 
+   @media (min-width: breakpoints.$bp-sm) {
+     .internal-text-slot {
+       padding: 24px 0px 24px 0px;
+     }
+   }
+ 
+   @media (min-width: breakpoints.$bp-md) {
+     .internal-text-slot {
+       padding: 40px 0px 40px 0px;
+     }
+   }
+ 
+   @media (min-width: breakpoints.$bp-lg) {
+     .internal-text-slot {
+       padding: 0px;
+     }
+   }
+ 
+   @media (max-width: breakpoints.$bp-lg) {
+     .full-bleed-grid {
+       grid-template-columns: 1fr;
+     }
+ 
+     .container-one.container-media {
+       grid-row-start: 1;
+       grid-column: 1 / auto;
+     }
+     .container-two.container-text {
+       grid-row-start: 2;
+       grid-column: 1 / auto;
+     }
+ 
+     .container-two.container-media {
+       grid-row-start: 1;
+       grid-column: 1 / auto;
+     }
+     .container-one.container-text {
+       grid-row-start: 2;
+       grid-column: 1 / auto;
+     }
+   }
+ }
+
