+++ src/components/eventSingleComponent/eventSingleComponent.test.ts
@@ -1,0 +1,10 @@
+import { EventSingleComponent } from '../../../dist/index';
+
+import { assert } from '@open-wc/testing';
+
+suite('kd-event-single-component', () => {
+  test('is defined', () => {
+    const el = document.createElement('kd-event-single-component');
+    assert.instanceOf(el, EventSingleComponent);
+  });
+});
