+++ src/components/eventSingleComponent/index.ts
@@ -1,2 +1,2 @@
 export { EventSingleComponent } from './eventSingleComponent';
-export { KdImage } from './image';
+//export { KdImage } from './image';
