+++ src/index.ts
@@ -1,2 +1,6 @@
 export { Eyebrow } from './components/eyebrow';
 export { AlternatingFeatureWithImage } from './components/alternatingFeatureWithImage';
+export { KdEventMultipleComponent } from './components/EventMultipleComponent';
+export { EventSingleBase } from './components/EventSingleBase';
+export { EventSingleComponent } from './components/eventSingleComponent';
+
